+++ lib/src/extensions/color_extension.dart
@@ -2,6 +2,6 @@ import 'dart:ui';
 
 extension ColorExtension on Color {
   String toHexString() {
-    return value.toRadixString(16);
+    return "#${value.toRadixString(16).padLeft(8, '0').toUpperCase()}";
   }
 }